[WEB-819] Move prescription form shared constants to own file
